requirements.txt
flask
scikit-learn
numpy
flask-cors
pillow